feat: Implementing card number verification when creating a new card
In this commit, I added a cardExists() method to the CardService class, which checks if a card already exists with the provided card number. If the card number is already in use, a 422 Unprocessable Entity response is returned to the user. If the card number is not in use, a new card is created and a 201 Created response is returned, with the new card information in the response body.